requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
idna==3.11
    # via requests
python-dotenv==1.2.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
urllib3==2.5.0
    # via requests
